package uses options now

+++ tasks/package.js
@@ -4,6 +4,10 @@ module.exports = function (grunt) {
 
     grunt.registerMultiTask('package', 'Handles the creation of minified JS/CSS files', function () {
         var self = this;
+        var options = this.options({
+            minify: false,
+            clear: []
+        });
 
         // concat the files
         var src = _.map(self.filesSrc, function (filepath) {
@@ -18,9 +22,11 @@ module.exports = function (grunt) {
         grunt.log.writeln('File ' + self.data.dest + ' created.');
 
         // clear unwanted files
-        _.each(self.data.clear || [],function(filepath){
-            grunt.file.delete(filepath);
-        });
+        if(_.isArray(options.clear)) {
+            _.each(options.clear || [],function(filepath){
+                grunt.file.delete(filepath);
+            });
+        }
     });
 
 };
